Delete deploy contract tool. Move mock to its own package as testonly with some helper functions
