+++ container/trie/BUILD.bazel
@@ -24,7 +24,7 @@ go_test(
         ":go_default_library",
         "//config/fieldparams:go_default_library",
         "//config/params:go_default_library",
-        "//contracts/deposit:go_default_library",
+        "//contracts/deposit/mock:go_default_library",
         "//crypto/hash:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
