+++ contracts/deposit/BUILD.bazel
@@ -5,8 +5,8 @@ go_library(
     srcs = [
         "contract.go",
         "deposit.go",
+        "helper.go",
         "logs.go",
-        "mock.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/contracts/deposit",
     visibility = ["//visibility:public"],
@@ -20,11 +20,8 @@ go_library(
         "@com_github_ethereum_go_ethereum//:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
-        "@com_github_ethereum_go_ethereum//accounts/abi/bind/backends:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
-        "@com_github_ethereum_go_ethereum//core:go_default_library",
         "@com_github_ethereum_go_ethereum//core/types:go_default_library",
-        "@com_github_ethereum_go_ethereum//crypto:go_default_library",
         "@com_github_ethereum_go_ethereum//event:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
     ],
@@ -43,6 +40,7 @@ go_test(
         "//beacon-chain/core/signing:go_default_library",
         "//config/params:go_default_library",
         "//container/trie:go_default_library",
+        "//contracts/deposit/mock:go_default_library",
         "//crypto/bls:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//runtime/interop:go_default_library",
