+++ testing/endtoend/components/BUILD.bazel
@@ -21,6 +21,7 @@ go_library(
         "//config/features:go_default_library",
         "//config/params:go_default_library",
         "//contracts/deposit:go_default_library",
+        "//contracts/deposit/mock:go_default_library",
         "//crypto/bls:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//runtime/interop:go_default_library",
