--- tools/deployContract/BUILD.bazel
@@ -1,27 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_binary")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["deployContract.go"],
-    importpath = "github.com/prysmaticlabs/prysm/tools/deployContract",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//contracts/deposit:go_default_library",
-        "//runtime/version:go_default_library",
-        "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
-        "@com_github_ethereum_go_ethereum//accounts/keystore:go_default_library",
-        "@com_github_ethereum_go_ethereum//crypto:go_default_library",
-        "@com_github_ethereum_go_ethereum//ethclient:go_default_library",
-        "@com_github_ethereum_go_ethereum//rpc:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_urfave_cli_v2//:go_default_library",
-        "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
-    ],
-)
-
-go_binary(
-    name = "deployContract",
-    embed = [":go_default_library"],
-    visibility = ["//visibility:public"],
-)
